requirements.txt
Flask==3.0.0
mysql-connector-python==8.4.0
undetected-chromedriver==3.5.5
urllib3==2.2.1
python-dotenv==1.0.1